MongoClientHelper to enable logging of messages sent to Mongo
Mongo client does not provide any logging of outgoing and incoming
request and responses. This is often required for test cases.

This can be done by registering a custom CommandListener. MongoClient
does not provide a direct way to configure it so we need to copy the
client construction logic to MongoClientHelper